git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/64bit-exp@1849 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ modules/image/consumer/image_consumer.cpp
@@ -89,7 +89,7 @@ public:
 		return info;
 	}
 
-	virtual size_t buffer_depth() const override
+	virtual int buffer_depth() const override
 	{
 		return 0;
 	}
